requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
requests
pytest
python-dotenv
numpy
soundfile